requirements.txt
requests==2.31.0
typing_extensions==4.2.0
pysqlite3-binary==0.5.2
